PWH v0.0.28] fix : material 미적용 오류 수정
