+++ Assets/ExternalAssets/QuickOutline/Resources/Materials/OutlineFill.mat
@@ -25,7 +25,7 @@ Material:
     m_TexEnvs: []
     m_Ints: []
     m_Floats:
-    - _OutlineWidth: 2
+    - _OutlineWidth: 10
     - _ZTest: 0
     m_Colors:
     - _OutlineColor: {r: 0.62264156, g: 2.6650329, b: 4, a: 1}
