+++ Assets/Prefabs/Food.prefab
@@ -17,6 +17,7 @@ GameObject:
   - component: {fileID: 9179529948731465183}
   - component: {fileID: 5883244226459090554}
   - component: {fileID: 6494852378264578925}
+  - component: {fileID: 2973207678377685388}
   m_Layer: 0
   m_Name: Food
   m_TagString: Untagged
@@ -339,10 +340,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6cdf13f29845c9e4cab53db221a30028, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  renderer: {fileID: 806663577967900737}
   eventHandler: {fileID: 6494852378264578925}
-  outlineFill_Prefab: {fileID: 2100000, guid: 0c4a0d014f61e154e8fc7327ead2c7c4, type: 2}
-  outlineMask_Prefab: {fileID: 2100000, guid: aac71b105816a6941afe786b14514338, type: 2}
+  outline: {fileID: 0}
 --- !u!114 &6494852378264578925
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -356,3 +355,21 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   interactor: {fileID: 0}
+--- !u!114 &2973207678377685388
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 789483719211524851}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5fea29bb7c508c244a1f805a5fd3fc4d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  outlineMode: 0
+  outlineColor: {r: 0, g: 1, b: 0.8771584, a: 1}
+  outlineWidth: 2
+  precomputeOutline: 0
+  bakeKeys: []
+  bakeValues: []
