+++ Assets/Scenes/SceneLobby.unity
@@ -11062,7 +11062,7 @@ Transform:
   m_GameObject: {fileID: 1322913536}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.6, y: 5.5, z: 20.5}
+  m_LocalPosition: {x: 0.6, y: 5.5, z: 17.800001}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
